make event stream common for viewport
