+++ web/src/components/Vnc/style.scss
